+++ build.gradle
@@ -46,7 +46,7 @@ application {
 }
 
 shadowJar {
-    archiveBaseName = "duke"
+    archiveBaseName = "taskie"
     archiveClassifier = null
 }
