--- src/main/java/duke/command/ListCommand.java
@@ -1,17 +1,0 @@
-package duke.command;
-
-import duke.storage.Storage;
-import duke.tasklist.TaskList;
-import duke.ui.Ui;
-
-public class ListCommand extends Command {
-    public ListCommand() {
-        super("list");
-    }
-
-    @Override
-    public void execute(TaskList tasks, Ui ui, Storage storage, StringBuilder response) {
-        response.append(ui.list(tasks.list()));
-        assert response.length() > 0; // response should not be empty
-    }
-}
