--- src/main/java/duke/command/MarkCommand.java
@@ -1,32 +1,0 @@
-package duke.command;
-
-import duke.storage.Storage;
-import duke.task.Task;
-import duke.tasklist.TaskList;
-import duke.ui.Ui;
-
-public class MarkCommand extends Command {
-    private int index;
-
-    public MarkCommand(int index) {
-        super("mark");
-        this.index = index;
-    }
-
-    @Override
-    public void execute(TaskList tasks, Ui ui, Storage storage, StringBuilder response) {
-        if (isOutOfBounds(tasks)) {
-            response.append("Invalid index, please try again.");
-            return;
-        }
-
-        Task task = tasks.mark(index);
-        response.append(ui.taskMarkedMessage(task));
-        assert response.length() > 0; // response should not be empty
-        storage.save(tasks.list());
-    }
-
-    private boolean isOutOfBounds(TaskList tasks) {
-        return index < 0 || index >= tasks.size();
-    }
-}
