--- src/main/java/duke/command/UnknownCommand.java
@@ -1,17 +1,0 @@
-package duke.command;
-
-import duke.storage.Storage;
-import duke.tasklist.TaskList;
-import duke.ui.Ui;
-
-public class UnknownCommand extends Command {
-    public UnknownCommand() {
-        super("unknown");
-    }
-
-    @Override
-    public void execute(TaskList tasks, Ui ui, Storage storage, StringBuilder response) {
-        response.append("Sorry, I don't understand that command :/");
-        assert response.length() > 0; // response should not be empty
-    }
-}
