--- src/main/java/duke/task/Event.java
@@ -1,19 +1,0 @@
-package duke.task;
-
-import java.time.LocalDate;
-import java.time.format.DateTimeFormatter;
-
-public class Event extends Task {
-    private LocalDate date;
-
-    public Event(String name, LocalDate date) {
-        super(name);
-        this.date = date;
-    }
-
-    @Override
-    public String toString() {
-        return "[E] " + super.toString() + " (at: "
-                + date.format(DateTimeFormatter.ofPattern("MMM d yyyy")) + ")";
-    }
-}
