--- src/main/java/duke/task/Todo.java
@@ -1,12 +1,0 @@
-package duke.task;
-
-public class Todo extends Task {
-    public Todo(String name) {
-        super(name);
-    }
-
-    @Override
-    public String toString() {
-        return "[T] " + super.toString();
-    }
-}
