+++ src/main/java/taskie/command/DeleteCommand.java
@@ -1,0 +1,32 @@
+package taskie.command;
+
+import taskie.storage.Storage;
+import taskie.task.Task;
+import taskie.tasklist.TaskList;
+import taskie.ui.Ui;
+
+public class DeleteCommand extends Command {
+    private int index;
+
+    public DeleteCommand(int index) {
+        super("delete");
+        this.index = index;
+    }
+
+    @Override
+    public void execute(TaskList tasks, Ui ui, Storage storage, StringBuilder response) {
+        if (isOutOfBounds(tasks)) {
+            response.append("Invalid index, please try again.");
+            return;
+        }
+            Task task = tasks.remove(index);
+            response.append(ui.taskDeleteMessage(task, tasks.size()));
+            assert response.length() > 0; // response should not be empty
+            storage.save(tasks.list());
+        }
+
+    private boolean isOutOfBounds(TaskList tasks) {
+        return index < 0 || index >= tasks.size();
+    }
+}
+
