+++ src/main/java/taskie/task/Deadline.java
@@ -1,0 +1,20 @@
+package taskie.task;
+
+import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
+
+public class Deadline extends Task {
+    private LocalDate date;
+
+    public Deadline(String name, LocalDate date) {
+        super(name);
+        this.date = date;
+    }
+
+    @Override
+    public String toString() {
+        return "[D] " + super.toString() + " (by: "
+                + date.format(DateTimeFormatter.ofPattern("MMM d yyyy")) + ")";
+    }
+}
+
