+++ src/main/java/taskie/task/Todo.java
@@ -1,0 +1,12 @@
+package taskie.task;
+
+public class Todo extends Task {
+    public Todo(String name) {
+        super(name);
+    }
+
+    @Override
+    public String toString() {
+        return "[T] " + super.toString();
+    }
+}
